Add structured exceptions with consistent messages
These messages are based on the existing pq_runtime exceptions and an
examination of current usage.  Rather than taking the usual free-form
message, they take the specific information that invariably makes up
the message, store those bits separately for the cases where recovery
is possible, and format the general message themselves for consistency.
With the exception of the ConnectionError, which may or may not carry
structured cause information, no freeform text is allowed.  In the cases
where additional explanation is truly needed, it can be logged.  Current usage
suggests such situations are rare.

Also remove useless conftest leftover from something or other.