Refactor SSHChannel.expect() for clarity.
I previously found the look sufficiently confusing that I thought it might
lead to dead code, although clearly that didn't make sense.  The way out
of the top section was to miss a deeply nested "if" with no else.  Otherwise
you raised or continued to the next iteration.  This splits the large
method into three and rearranges the calls to eliminate the need
for the continue.

ReviewID: 98605