Push common channel code to base class.
While working on the KVM console serial channel I noticed a good bit of
duplicate code- some common to all three channels, some just commmon to
ssh and serial.  This pushes those up to the base class, and adds
an abstract method for a useful helper to support them.

There is more consolidation that can be done in the unit tests, but
I would like to wait until the serial channel is done before tackling
that.

ReviewID: 98686