Partial libvirt channel for pq-cmdline
Implementation of our channel interface for libvirt serial connections.
Note that this is not actually specific to KVM (in theory).
This is not actually usable yet as it does not have the receive_all,
expect, or initial login code present.  This is an easier chunk to
start with for a review.

Also standardize on the logging.* functions instead of instantiating
individual logs.

DesignURL: http://rtd.lab.nbttech.com/docs/qa-design-review/en/latest/handrews/pq-console-cli.html
ReviewID: 98849