[PQ] Fix Cisco router login issue
Attempt to login the Cisco router was failing because of -
E       UnicodeDecodeError: 'ascii' codec can't decode byte 0xff in
position 0: ordinal not in range(128)

This was happening because the password was being sent as a unicode and
not encoded as an ascii . The fix is simple one - make sure that the password is
encoded as a ascii. I didnt hit the issue with username (because this part of code
never gets executed during router login) but I am not taking any chances.

ReviewID: 106418
Review By: jheimann
Approved By: jheimann